Extracted git functionality into a new utility module.
