+++ src/commands/clone.js
@@ -3,7 +3,6 @@ const { join } = require( 'path' );
 const { tmpdir, EOL } = require( 'os' );
 
 const chalk = require( 'chalk' );
-const inquirer = require( 'inquirer' );
 const fsExtra = require( 'fs-extra' );
 
 const { replaceLinks } = require( '../utils/make-link' );
@@ -33,7 +32,6 @@ exports.builder = function( yargs ) {
 	yargs.option( 'b', {
 		alias: 'branch',
 		description: 'Branch name to checkout',
-		default: 'master',
 		type: 'string',
 	} );
 
@@ -46,13 +44,11 @@ exports.builder = function( yargs ) {
 };
 
 exports.handler = makeCommand( async ( { url, branch, config, verbose } ) => {
-	const git = require( 'nodegit' ); // nodegit must be required here
-
 	const tempDir = mkdtempSync( join( tmpdir(), 'wpld-' ) );
 	const spinner = ! verbose ? makeSpinner() : undefined;
 
 	// clone repository
-	await makeGitClone( spinner, chalk, git, inquirer )( tempDir, url, branch );
+	await makeGitClone( spinner )( tempDir, url, branch );
 	// read configuration from the config file in the repo if it exists
 	const configuration = await makePullConfig( spinner )( tempDir, config );
 	// create environment
